feat: implementação da deleção de movimentações de saidas

+++ Backend/CeramicaCanelas.Application/Features/Movimentacoes-ES/Entradas/Command/DeleteMovEntradasProductsCommand/DeleteMovEntradasProductsCommandHandler.cs
@@ -13,12 +13,14 @@ namespace CeramicaCanelas.Application.Features.Movimentacoes_ES.Entradas.Command
     public class DeleteMovEntradasProductsCommandHandler : IRequestHandler<DeleteMovEntradasProductsCommand, Unit>
     {
         private readonly IMovEntryProductsRepository _repository;
+        private readonly IProductRepository _productRepository;
         private readonly ILogged _logged;
 
-        public DeleteMovEntradasProductsCommandHandler(IMovEntryProductsRepository repository, ILogged logged)
+        public DeleteMovEntradasProductsCommandHandler(IMovEntryProductsRepository repository, ILogged logged, IProductRepository productRepository)
         {
             _repository = repository;
             _logged = logged;
+            _productRepository = productRepository;
         }
 
         public async Task<Unit> Handle(DeleteMovEntradasProductsCommand command, CancellationToken cancellationToken)
@@ -35,6 +37,19 @@ namespace CeramicaCanelas.Application.Features.Movimentacoes_ES.Entradas.Command
             if (movimentacaoES == null)
                 throw new BadRequestException("Movimentação de entrada não encontrada.");
 
+            // Atualiza o estoque do produto
+            var product = await _productRepository.GetProductByIdAsync(movimentacaoES.ProductId);
+            if (product == null)
+            {
+                throw new BadRequestException("Produto não encontrado.");
+            }
+            //Desatalizando a movimentação anterior
+            product.StockCurrent -= movimentacaoES.Quantity;
+            product.ValueTotal -= movimentacaoES.UnitPrice * movimentacaoES.Quantity;
+            product.ModifiedOn = DateTime.UtcNow;
+            await _productRepository.Update(product);
+
+            // Exclui a movimentação de entrada
             await _repository.Delete(movimentacaoES);
 
             return Unit.Value;
